Move Kestrel helper to RealStartup and add websocket end point.
